requirements.txt
Django==4.1
djangorestframework==3.13.1
django_rest_passwordreset==1.2.1
drf_yasg==1.21.3
django-cors-headers==3.13.0
PyJWT==2.4.0
pytest==7.1.2
requests==2.28.1
sentry_sdk==1.9.2
six==1.16.0
stripe==4.0.2
xhtml2pdf==0.2.8
xlwt==1.3.0
django-rest-knox==4.2.0
django-ckeditor==6.4.2
dj-stripe==2.6.2
psycopg2-binary==2.9.3
djangorestframework-simplejwt==5.2.0
django-filter==22.1
whitenoise==6.2.0
pytest-django==4.5.2
Faker==13.15.1
factory-boy==3.2.1
gunicorn==20.1.0
channels==3.0.5
channels_redis==3.4.1
asgiref==3.5.2
